SW-4489 Refactor the entity filters so thwe filter config for the filters specific to each project entity are passed in
